Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.3.0
chromadb==0.4.18
openai==1.3.7
google-generativeai==0.3.2
PyMuPDF==1.23.8
cryptography==41.0.7
httpx==0.24.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4